fix: panic when repository has already been cloned
In the case that a user runs the program with the same name argument
inside the same folder, the program would panic due to it not checking
if a folder with the same name already exists. The program now exits
with an exit code of 1 instead and prints out a message on stderr.

Fixes #2