fix(cli): cargo subcommand not working properly
Cargo invokes the subcommand binary by calling it and passing the binary
name as well as the subcommand that it represents as arguments. Those
have to be stripped away before `clap` can proceed and parse the
arguments.

The change resulted in the need to create two binaries inside the `bin`
folder and move the crate to a library type. The API inside the library
should under no circumstances be assumed to be stable. Semantic
versioning is only applied to the overall project and the library should
be handled as internal API only.

Fixes #13