feat: move/rename files
The `files` field inside of the `jen.toml` file now takes two types of
values inside the array. A simple string that points to a file.
Additionally, an inline table with the values `from` and `to` can be
used to move files inside the template repository during the feature
resolve phase.

Moving can be used to structure template repositories and make their
development a little bit more bearable. Renaming files can be used if
different features want to a file with the same name.

Closes #32